requirements.txt
asgiref
Django
django-cors-headers
django-environ
django-filter
djangorestframework
environ
Markdown
pillow
sqlparse
tzdata
